use tid instead of pid on linux

+++ src/format.rs
@@ -201,7 +201,7 @@ fn format_datetime(
 }
 
 pub(crate) struct FormatProcessData<'a> {
-    pub(crate) pid: u32,
+    pub(crate) pid: i64,
     pub(crate) thread_name: Option<&'a str>,
     pub(crate) with_thread_names: bool,
     pub(crate) metadata: &'a Metadata<'a>,
